Add files via upload
The purpose of this program is to calculate the final grade for the course Programming Methods 
Methods I, by periodic evaluation, on a scale of 0 to 20.
Initially, I opened the program and defined the following variables, of type double:
gradeP1, grade for the first project; gradeP2, grade for the second project; gradeF1, grade for the 
grade; F2, the second grade; and S, the final sum.
I asked you, using a printf, which prints a message on the screen, to define “Final grade of 
final grade of Programming Methods 1” as the title. I then asked you to 
the grade for the first project, on a scale of 0 to 2, and asked it to read it, 
using a scanf. Next, I asked him to enter the grade for the second project, on a scale of 0 to 4. 
project, on a scale of 0 to 4, and asked it to be read out. 
Then I asked them to enter the grade for the first and second attendance, respectively. 
frequency, respectively. Since these are lectures, the grade entered must be written on a scale of 0 to 4. 
must be written on a scale of 0 to 20. 
If the scales are not respected, the message “Data error” will appear and the user will have to 
user will have to start again, respecting the defined limits.
I defined the sum of the final grade and asked for the count to be carried out. To do this, I had to define 
another integer variable, S1. And I used the round function which rounds off the final grade.
Finally, I asked the program to write down what the final grade corresponded to. 
The scale is as follows: from 0 to 9 Insufficient, from 10 to 13 Sufficient, from 14 to 16 Good 
and from 17 to 20 Very Good.